Servo controller

+++ capybara/src/muskrat.rs
@@ -1,6 +1,6 @@
 use tokio::sync::mpsc;
 
-pub async fn run_muskrat(mut set_angle_rx: mpsc::Receiver<u8>) {
+pub async fn run_muskrat(mut set_angle_rx: mpsc::Receiver<f64>) {
     while let Some(angle) = set_angle_rx.recv().await {
         println!("New angle = {}", angle);
     }
